requirements.txt
awsebcli==3.5.6
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
botocore==1.3.8
bs4==0.0.1
cement==2.4.0
colorama==0.3.3
DateTime==4.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.5
django-filter==0.13.0
django-toolbelt==0.0.1
djangorestframework==3.3.3
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
Flask==0.10.1
gunicorn==19.4.5
heroku==0.1.4
itsdangerous==0.24
Jinja==1.2
Jinja2==2.8
jmespath==0.9.0
Markdown==2.6.6
MarkupSafe==0.23
pathspec==0.3.3
pdb==0.1
psycopg2==2.6.1
python-dateutil==1.5
python-gnupg==0.3.8
pytz==2016.4
PyYAML==3.11
requests==2.6.2
six==1.10.0
static3==0.6.1
texttable==0.8.4
websocket-client==0.34.0
Werkzeug==0.11.3
whitenoise==3.0
zope.interface==4.1.3
